+++ src/components/JobCard/JobCardCommunication.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Job } from '../../types/Job';
 
 interface JobCardCommunicationProps {